Add simple joystick example
